exported validation
